% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coins.R
\name{autoCorrCoin}
\alias{autoCorrCoin}
\title{Create An Autocorrelated Coin}
\usage{
autoCorrCoin(success = "H", failure = "T", initial_prob = 0.5,
  p_given1 = 0.5, p_given0 = 0.5)
}
\arguments{
\item{success}{Denotes "success" in the Bernoulli Trial. Default is "H".}

\item{failure}{Denotes "failure" in the Bernoulli Trial. Default is "F".}

\item{initial_prob}{The initial probability of "success". Must be numeric and be between 0 and 1. Default is 0.5.}

\item{p_given1}{The probability of "success" given the previous trial was a "success". Must be numeric and be between 0 and 1. Default is 0.5.}

\item{p_given0}{The probability of "success" given the previous trial was a "failure". Must be numeric and be between 0 and 1. Default is 0.5.}
}
\description{
Creates an autocorrelated coin
}
\author{
Justin Strate
}
