% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coins.R
\name{coin}
\alias{coin}
\title{Create New Coin}
\usage{
coin(success = "H", failure = "T", prob = 0.5)
}
\arguments{
\item{success}{Denotes "success" in the Bernoulli Trial. Default is "H"}

\item{failure}{Denotes "failure" in the Bernoulli Trial. Default is "F"}

\item{prob}{The probability of "success". Must be numeric and be between 0 and 1. Default is 0.5}
}
\description{
Creates a new coin or Bernoulli Trial
}
\author{
Justin Strate
}
