% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coin_toss}
\alias{coin_toss}
\title{Fictional Cointosses}
\format{An object of class \code{factor} of length 31.}
\usage{
coin_toss
}
\description{
A factor vector represening fictional cointosses
}
\keyword{datasets}
