% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coins.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{foo}
\title{Convert \code{data.frame} to \code{list}.}
\usage{
foo(x)

lhs \%>\% rhs
}
\arguments{
\item{x}{A \code{data.frame} object.}
}
\description{
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\examples{
my_result <- foo(iris)

}
\keyword{internal}
