% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coins.R
\name{print.tossedCoin}
\alias{print.tossedCoin}
\title{Print Tossed Coin}
\usage{
\method{print}{tossedCoin}(x, ...)
}
\arguments{
\item{x}{the object to be printed}

\item{...}{other arguments that may be supplied}
}
\description{
prints the trial of the tossed coin
}
