% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinAnalytics.R
\name{streaks.default}
\alias{streaks.default}
\title{streak}
\usage{
\method{streaks}{default}(x, ...)
}
\arguments{
\item{x}{a binary vector of successes and failures}

\item{...}{additional parameters}
}
\value{
an integer vector that records the number of success
}
\description{
counts streaks
}
\author{
Justin Strate
}
