% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinAnalytics.R
\name{streaks}
\alias{streaks}
\title{Streaks}
\usage{
streaks(...)
}
\arguments{
\item{...}{Additional parameters that may be applied}
}
\value{
An integer vector that records the number of streaks of successes
}
\description{
Counts the streaks
}
\author{
Justin Strate
}
