% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinAnalytics.R
\name{streaks.tossedCoin}
\alias{streaks.tossedCoin}
\title{streaks.tossedCoin}
\usage{
\method{streaks}{tossedCoin}(mycoin, ...)
}
\arguments{
\item{mycoin}{a coin object}

\item{...}{Additional parameters that may be applied}
}
\description{
coints the streaks for a coin that has been tossed
}
