% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tosses.R
\name{toss.autoCorrCoin}
\alias{toss.autoCorrCoin}
\title{Toss An Autocorrelated Coin}
\usage{
\method{toss}{autoCorrCoin}(mycoin, n)
}
\arguments{
\item{mycoin}{An Autocorrelated Coin}

\item{n}{The number of times to toss the autocorrelated coin}
}
\description{
Creates an autocorrelated coin
}
\author{
Justin Strate
}
