% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tosses.R
\name{toss.coin}
\alias{toss.coin}
\title{Toss a Coin}
\usage{
\method{toss}{coin}(mycoin, n, ...)
}
\arguments{
\item{mycoin}{an object of class coin}

\item{n}{the number of times to toss the coin}

\item{...}{additional parameters that may be supplied.}
}
\description{
Toss a coin
}
