% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tosses.R
\name{tossed_coin.character}
\alias{tossed_coin.character}
\title{Creating a Tossed Coing from a Vector}
\usage{
\method{tossed_coin}{character}(vec, success, failure, prob = NULL, ...)
}
\arguments{
\item{vec}{a character vector reprsenting the trial}

\item{success}{denotes "success" in the trial}

\item{failure}{denotes "failure" in the trial}

\item{prob}{Represents the probability of "Success". Default is NULL}

\item{...}{other arguments that may be supplied}
}
\description{
Creates a tossed coin from either a characeter or factor vector of two levels
}
