% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coins.R
\name{print.coin}
\alias{print.coin}
\title{Print Coin}
\usage{
\method{print}{coin}(coin, ...)
}
\arguments{
\item{coin}{a coin object}

\item{...}{Other arguments that may be supplied}
}
\description{
prints a coin
}
\author{
Justin Strate
}
