% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tosses.R
\name{toss}
\alias{toss}
\title{Coin Toss}
\usage{
toss(...)
}
\arguments{
\item{...}{additional parameters that may be supplied.}
}
\description{
A generic gunction to simulate the tossing of a coin
}
\author{
Justin Strate
}
